# lib/api-zod/src/generated/types/registrationConfirmation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Treinão Encontro das Águas — 2ª Edição
 * OpenAPI spec version: 0.1.0
 */

export interface RegistrationConfirmation {
  id: number;
  registrationCode: string;
  fullName: string;
  routeName: string;
  eventName: string;
  eventDate: string;
  eventTime: string;
  location: string;
  createdAt: string;
}