# lib/api-zod/src/generated/types/registrationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Treinão Encontro das Águas — 2ª Edição
 * OpenAPI spec version: 0.1.0
 */
import type { RegistrationInputSex } from './registrationInputSex';
import type { RegistrationInputShirtSize } from './registrationInputShirtSize';

export interface RegistrationInput {
  eventId: number;
  routeId: number;
  /** @minLength 2 */
  fullName: string;
  /**
     * @minLength 11
     * @maxLength 14
     * @pattern ^\d{3}\.?\d{3}\.?\d{3}-?\d{2}$
     */
  cpf: string;
  /** @minLength 8 */
  whatsapp: string;
  /** @nullable */
  email?: string | null;
  birthDate: Date;
  sex: RegistrationInputSex;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  neighborhood?: string | null;
  /** @nullable */
  team?: string | null;
  /** @nullable */
  shirtSize?: RegistrationInputShirtSize;
  /** @nullable */
  medicalNotes?: string | null;
  /** @nullable */
  emergencyContactName?: string | null;
  /** @nullable */
  emergencyContactWhatsapp?: string | null;
  acceptedTerms: boolean;
  /** @nullable */
  termsVersion?: string | null;
}